[GSoC Milestone 1] Migrate resource calls from RESTEasy to GitLab4J-API (#1501)
* adding gitlab4j dependency

* temporarily adapting causedata

* migrating connection

* removing models and enums

* migrating gitlab api

* adapting models

* migrating publishers

* migrating services

* migrating utils

* migrating handlers

* adapting gitlab api tests

* migrating workflow

* adapting env contributor test

* migrating test helpers

* migrating tests

* Debug

* Remove skip unit tests config in pom.xml

* Put back mistakenly removed code

* remove unnecessary models

* re-enabling tests

* Passing CI build

* patch according to code review

* removing resteasy titbits

* minor test patches

* Revert previous spotbug exclusions no longer needed

* enabling tests

* Patch PendingBuildsHandlerTest test cases

* fixing client-builderid not found problem in tests

* Remove unneeded transitive dependencies

* fixing angry jenkins on testing connection

* Fix Mockito tests not working due to equals() not implemented

* minor fix and enabling tests

* minor test patches

* fixing some connection tests

* Disable tests and apply Spotless formatting to files

* minor patch in tests


---------

+++ src/main/java/com/dabsquared/gitlabjenkins/gitlab/hook/model/PipelineHook.java
@@ -1,3 +1,9 @@
+/* Note to Reviewers :
+ * Will be removed.
+ */
+/*
+ * TODO remove this class because in milestone 2 PipelineEvent object will be used instead of PipelineHook
+ */
 package com.dabsquared.gitlabjenkins.gitlab.hook.model;
 
 import java.util.List;
@@ -13,16 +19,16 @@ import org.apache.commons.lang.builder.ToStringBuilder;
 public class PipelineHook extends WebHook {
 
     private User user;
-    public Integer projectId;
+    public Long projectId;
     private List<Commit> commits;
     private Project project;
     private PipelineEventObjectAttributes objectAttributes;
 
-    public Integer getProjectId() {
+    public Long getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(Integer projectId) {
+    public void setProjectId(Long projectId) {
         this.projectId = projectId;
     }
