+++ src/main/java/com/dabsquared/gitlabjenkins/gitlab/hook/model/PushHook.java
@@ -20,7 +20,7 @@ public class PushHook extends WebHook {
     private String userUsername;
     private String userEmail;
     private String userAvatar;
-    private Integer projectId;
+    private Long projectId;
     private Project project;
     private List<Commit> commits;
     private Integer totalCommitsCount;
@@ -89,11 +89,11 @@ public class PushHook extends WebHook {
         this.userAvatar = userAvatar;
     }
 
-    public Integer getProjectId() {
+    public Long getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(Integer projectId) {
+    public void setProjectId(Long projectId) {
         this.projectId = projectId;
     }
